Fix identity

+++ test/segment_tree.test.cpp
@@ -5,7 +5,7 @@
 
 int main() {
     int n, q; std::cin >> n >> q;
-    SegmentTree seg(std::vector<int>(n, (1LL << 31) - 1));
+    SegmentTree seg(std::vector<long long>(n, (1LL << 31) - 1));
     for (int i = 0; i < q; i++) {
         int com, x, y; std::cin >> com >> x >> y;
         if (com) std::cout << seg.getval(x, y + 1) << std::endl;
